# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: 智慧物业管理系统 SaaS 多租户版 API
  version: 1.1.1
  description: |
    本接口契约用于约束 Codex 开发。实际开发不得擅自更改路径、核心字段、状态枚举和权限标识。
    所有业务接口默认返回 ApiResponse；列表接口 data 使用 PageResult。
servers:
  - url: https://api.example.com
    description: 生产环境示例
  - url: http://localhost:8080
    description: 本地开发环境
security:
  - bearerAuth: []
...
paths:
...
  /api/service/workorders:
    get:
      tags: [Service]
      summary: 工单列表
      x-permission: service:workorder:list
      parameters:
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/WorkOrderStatus' }
        - name: orderType
          in: query
          schema: { type: string, enum: [REPAIR, COMPLAINT, SUGGESTION, HOUSEKEEPING, RENOVATION] }
        - name: handlerUserId
          in: query
          schema: { type: integer, format: int64 }
        - name: memberId
          in: query
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
    post:
      tags: [Service]
      summary: 后台创建工单
      x-permission: service:workorder:create
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderCreateRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/service/workorders/{workOrderId}:
    get:
      tags: [Service]
      summary: 工单详情
      x-permission: service:workorder:view
      parameters:
        - $ref: '#/components/parameters/WorkOrderId'
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/service/workorders/{workOrderId}/dispatch:
    put:
      tags: [Service]
      summary: 工单派单
      x-permission: service:workorder:dispatch
      parameters:
        - $ref: '#/components/parameters/WorkOrderId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderDispatchRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/service/workorders/{workOrderId}/accept:
    put:
      tags: [Service]
      summary: 工单受理
      x-permission: service:workorder:accept
      parameters:
        - $ref: '#/components/parameters/WorkOrderId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderActionRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/service/workorders/{workOrderId}/reject:
    put:
      tags: [Service]
      summary: 工单驳回
      x-permission: service:workorder:reject
      parameters:
        - $ref: '#/components/parameters/WorkOrderId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderActionRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/service/workorders/{workOrderId}/start:
    put:
      tags: [Service]
      summary: 开始处理工单
      x-permission: service:workorder:process
      parameters:
        - $ref: '#/components/parameters/WorkOrderId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderActionRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/service/workorders/{workOrderId}/submit-result:
    put:
      tags: [Service]
      summary: 提交处理结果
      x-permission: service:workorder:process
      parameters:
        - $ref: '#/components/parameters/WorkOrderId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderActionRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/service/workorders/{workOrderId}/hang-up:
    put:
      tags: [Service]
      summary: 挂起工单
      x-permission: service:workorder:process
      parameters:
        - $ref: '#/components/parameters/WorkOrderId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderActionRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/service/workorders/{workOrderId}/resume:
    put:
      tags: [Service]
      summary: 恢复工单
      x-permission: service:workorder:process
      parameters:
        - $ref: '#/components/parameters/WorkOrderId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderActionRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/service/workorders/{workOrderId}/revisit:
    put:
      tags: [Service]
      summary: 工单回访完成
      x-permission: service:workorder:revisit
      parameters:
        - $ref: '#/components/parameters/WorkOrderId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderActionRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/service/workorders/{workOrderId}/rework:
    put:
      tags: [Service]
      summary: 工单返工
      x-permission: service:workorder:revisit
      parameters:
        - $ref: '#/components/parameters/WorkOrderId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderActionRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/service/workorders/{workOrderId}/cancel:
    put:
      tags: [Service]
      summary: 取消工单
      x-permission: service:workorder:cancel
      parameters:
        - $ref: '#/components/parameters/WorkOrderId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderActionRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/service/workorders/sla/mark-overdue:
    post:
      tags: [Service]
      summary: 标记 SLA 超时工单
      x-permission: service:workorder:sla
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/service/complaints:
    get:
      tags: [Service]
      summary: 投诉建议列表
      x-permission: service:complaint:list
      parameters:
        - name: projectId
          in: query
          schema: { type: integer, format: int64 }
        - name: status
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
...
  /api/app/workorders:
    post:
      tags: [App]
      summary: 小程序提交报修/投诉
      x-permission: app:workorder:create
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderCreateRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
    get:
      tags: [App]
      summary: 我的工单列表
      x-permission: app:workorder:list
      parameters:
        - name: projectId
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - name: memberId
          in: query
          required: true
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/PageNo'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/PageResponse' }
  /api/app/workorders/{workOrderId}:
    get:
      tags: [App]
      summary: 小程序工单详情
      x-permission: app:workorder:view
      parameters:
        - $ref: '#/components/parameters/WorkOrderId'
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/app/workorders/{workOrderId}/evaluate:
    post:
      tags: [App]
      summary: 小程序评价工单
      x-permission: app:workorder:evaluate
      parameters:
        - $ref: '#/components/parameters/WorkOrderId'
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderEvaluateRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
  /api/app/complaints:
    post:
      tags: [App]
      summary: 小程序提交投诉建议
      x-permission: app:complaint:create
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WorkOrderCreateRequest' }
      responses:
        '200': { $ref: '#/components/responses/ApiResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageNo:
      name: pageNo
      in: query
      schema: { type: integer, default: 1, minimum: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
...
    WorkOrderId:
      name: workOrderId
      in: path
      required: true
      schema: { type: integer, format: int64 }
...
  responses:
    ApiResponse:
      description: OK
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiResponse' }
    PageResponse:
      description: OK
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  data: { $ref: '#/components/schemas/PageResult' }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: integer, example: 0 }
        message: { type: string, example: success }
        data: { type: object, nullable: true }
        traceId: { type: string }
...
    PageResult:
      type: object
      properties:
        records:
          type: array
          items: { type: object }
        pageNo: { type: integer }
        pageSize: { type: integer }
        total: { type: integer, format: int64 }
...
    WorkOrderStatus:
      type: string
      enum: [SUBMITTED, ACCEPTED, DISPATCHED, PROCESSING, HANG_UP, WAIT_CONFIRM, COMPLETED, EVALUATED, CANCELLED, REJECTED, TRANSFERRED, REWORK]
...
    WorkOrderCreateRequest:
      type: object
      required: [projectId, orderType, title]
      properties:
        projectId: { type: integer, format: int64 }
        memberId: { type: integer, format: int64, nullable: true }
        houseId: { type: integer, format: int64 }
        orderType: { type: string, enum: [REPAIR, COMPLAINT, SUGGESTION, HOUSEKEEPING, RENOVATION] }
        title: { type: string }
        description: { type: string }
        location: { type: string }
        imageFileIds: { type: string }
        priority: { type: string, enum: [LOW, NORMAL, HIGH, URGENT] }
    WorkOrderActionRequest:
      type: object
      properties:
        content: { type: string }
        imageFileIds: { type: string }
    WorkOrderDispatchRequest:
      type: object
      required: [handlerUserId]
      properties:
        handlerUserId: { type: integer, format: int64 }
        content: { type: string }
    WorkOrderEvaluateRequest:
      type: object
      required: [memberId, score]
      properties:
        memberId: { type: integer, format: int64 }
        score: { type: integer, minimum: 1, maximum: 5 }
        content: { type: string }
...